+++ src/MachinaGrasshopper/Actions/ExternalAxis.cs
@@ -42,12 +42,12 @@ namespace MachinaGrasshopper.Actions
             // Absolute
             mpManager.AddComponentNames(false, "ExternalAxisTo", "ExternalAxisTo", "Set the values of one of the robot's external axes.");
             mpManager.AddParameter(false, typeof(Param_Integer), "AxisNumber", "EAid", "Axis number from 1 to 6.", GH_ParamAccess.item);
-            mpManager.AddParameter(false, typeof(Param_Number), "Increment", "inc", "Increment value in mm or degrees.", GH_ParamAccess.item);
+            mpManager.AddParameter(false, typeof(Param_Number), "Value", "v", "Increment value in mm or degrees.", GH_ParamAccess.item);
 
             // Relative
             mpManager.AddComponentNames(true, "ExternalAxis", "ExternalAxis", "Increase the values of one of the robot's external axes.");
             mpManager.AddParameter(true, typeof(Param_Integer), "AxisNumber", "EAid", "Axis number from 1 to 6.", GH_ParamAccess.item);
-            mpManager.AddParameter(true, typeof(Param_Number), "Value", "v", "New value in mm or degrees.", GH_ParamAccess.item);
+            mpManager.AddParameter(true, typeof(Param_Number), "Increment", "inc", "New value in mm or degrees.", GH_ParamAccess.item);
         }
 
         protected override void RegisterOutputParams(GH_OutputParamManager pManager)
@@ -60,12 +60,12 @@ namespace MachinaGrasshopper.Actions
             int axisNumber = 1;
             double val = 0;
 
-            if (DA.GetData(0, ref axisNumber)) return;
-            if (DA.GetData(1, ref val)) return;
+            if (!DA.GetData(0, ref axisNumber)) return;
+            if (!DA.GetData(1, ref val)) return;
 
             if (axisNumber < 1 || axisNumber > 6)
             {
-                AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "AxisNumber must be 1 to 6");
+                AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "AxisNumber must be between 1 and 6");
             }
 
             DA.SetData(0, new ActionExternalAxis(axisNumber, val, this.Relative));